api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
jinja2==3.1.4
python-multipart==0.0.9